Fixed x-axis overflow on tables + added unused utility function for checking overflow. (#48)
* Tidy in Overview.razor

Removed empty lines. Fixed typos in comments. Simplified code.

* WIP Overflow handling + overflow dropdown button styling

Also added a bit of dummy data to simulate overflow, as well as fixing some indentation in _Host.cshtml

* WIP

* WIP overflow checker

Currently works in .razor files, but not in Radzen.

* Resize the side-scrollbar

* Temp fix for x-axis overflow on tables

Also added OverflowChecker.js utility function to detect overflow, but it doesn't work the same in .razor as in Radzen. Unsure how to implement, hence the temp fix.

* Update ReconciliationTable.razor

* Update .gitignore

* Fix manager name overflowing